+++ animate.py
@@ -1,11 +1,10 @@
 import photomosaic as pm
-import pygame
 import Image
 import sys
 import random
 import operator
 from sql_image_pool import SqlImagePool
-pygame.init()
+from gui_utils import MosaicGUI
 
 DEPTH = 2
 
@@ -23,15 +22,6 @@ match_sort = color_sort
 
 locs = [[0,0], [1,0], [0,1], [1,1]]
 
-def topygame(img):
-    return pygame.image.frombuffer(img.tostring(), img.size, "RGB")
-
-def draw(img, box):
-    screen.blit(topygame(img), box)
-
-def draw_scaled(filename, x,y,w,h):
-    screen.blit(pygame.transform.scale(pygame.image.load(filename), (w, h)), (x,y))
-
 infile = sys.argv[1]
 database = sys.argv[2]
 tune = '-tune' in sys.argv
@@ -42,9 +32,10 @@ p = pm.Photomosaic(infile, pool, tuning=tune)
 W,H = p.orig_img.size
 Hx = 0
 Wx = W
-screen = pygame.display.set_mode((W+Wx,H+Hx))
-draw(p.orig_img, (0,0))
-pygame.display.flip()
+gui = MosaicGUI((W+Wx,H+Hx))
+
+gui.img(p.orig_img, (0,0))
+gui.draw()
 
 p.partition_tiles(10, depth=DEPTH)
 
@@ -53,9 +44,9 @@ for tile in sorted(p.tiles, key=analyze_sort):
     w,h = tile.size
     for (x,y), color in zip(locs, tile.rgb):
         rect = (Wx+tx + w*x/2,Hx+ty+h*y/2,w/2,h/2)
-        pygame.draw.rect(screen, color, rect)
-        pygame.draw.rect(screen, (0,0,0), rect, 1)
-        pygame.display.flip()
+        gui.rectangle(color, rect)
+        gui.rectangle((0,0,0), rect, 1)
+        gui.draw()
 
 try:
     for tile in sorted(p.tiles, key=match_sort):
@@ -64,12 +55,10 @@ try:
         p.match_one(tile)    
         tx,ty = tile.get_position(tile.size)
         w,h = tile.size
-        draw_scaled(tile.match[4], tx+Wx, ty+Hx, w, h)
-        pygame.display.flip()
+        gui.scaled_img(tile.match[4], tx+Wx, ty+Hx, w, h)
+        gui.draw()
 
 finally:
     pool.close()
-while True:
-    for event in pygame.event.get():
-        if event.type == pygame.QUIT: sys.exit()
+gui.wait_for_close()
